requirements.txt
streamlit>=1.30.0
openai>=1.0.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0